+++ src/Console/ModelsCommand.php
@@ -1291,7 +1291,8 @@ class ModelsCommand extends Command
         $methodReflection = $castReflection->getMethod('get');
 
         return $this->getReturnTypeFromReflection($methodReflection) ??
-            $this->getReturnTypeFromDocBlock($methodReflection, $reflection);
+            $this->getReturnTypeFromDocBlock($methodReflection, $reflection) ??
+            'mixed';
     }
 
     /**